+++ app/routes/components/Avatar/changelog.codegen.ts
@@ -6,4 +6,6 @@ export const changelog = ''
   + '- 2025-10-03: Add `Avatar.Img` for image support. Update styles to ensure images\n'
   + '  fit within the avatar circle.\n'
   + '- 2025-10-11: Using type `ColorName` for color prop.\n'
+  + '- 2025-10-21: Update avatar background color calculation for better contrast.\n'
+  + '  And add a border.\n'
   ;