+++ app/routes/components/Avatar/demos/Demo02.tsx
@@ -1,0 +1,31 @@
+import { Avatar } from "src/components/Avatar";
+import { Flex } from "@/components/Flex";
+
+export default function Demo() {
+  return (
+    <Flex dir="column" alignItems="center" gap="1rem">
+      <Flex dir="row" alignItems="center" gap="1rem">
+        <Avatar color="blue" size="2.5rem">
+          PZ
+        </Avatar>
+        <Avatar color="red" size="2.5rem">
+          PZ
+        </Avatar>
+        <Avatar color="green" size="2.5rem">
+          PZ
+        </Avatar>
+      </Flex>
+      <Flex dir="row" alignItems="center" gap="1rem">
+        <Avatar size="2.5rem">
+          <Avatar.Img
+            src="https://github.com/peterlitszo.png"
+            alt="Peterlits Zo"
+          />
+        </Avatar>
+        <Avatar size="2.5rem">
+          <Avatar.Img src="https://github.com/shadcn.png" alt="Shadcn" />
+        </Avatar>
+      </Flex>
+    </Flex>
+  );
+}