+++ src/components/Avatar/index.ts
@@ -1,3 +1,3 @@
 // Avatar from copy-ui @ 2025-10-13.
 
-export { Avatar } from "./Avatar";
+export { Avatar } from "./avatar";