+++ src/components/Avatar/CHANGELOG.md
@@ -2,3 +2,5 @@
 - 2025-10-03: Add `Avatar.Img` for image support. Update styles to ensure images
   fit within the avatar circle.
 - 2025-10-11: Using type `ColorName` for color prop.
+- 2025-10-21: Update avatar background color calculation for better contrast.
+  And add a border.